Changed: Moved benchmark-only functions in standard transform to 'bench' module.
