+++ projects/dxt-lossless-transform-bc1/benches/transform_standard/avx512.rs
@@ -1,5 +1,5 @@
 use criterion::{black_box, BenchmarkId};
-use dxt_lossless_transform_bc1::transforms::standard::transform::*;
+use dxt_lossless_transform_bc1::transforms::standard::transform::bench::*;
 use safe_allocator_api::RawAlloc;
 
 fn bench_avx512_permute_unroll_2(
