+++ projects/dxt-lossless-transform-bc1/benches/untransform_standard/avx2.rs
@@ -1,5 +1,7 @@
 use criterion::*;
-use dxt_lossless_transform_bc1::transforms::standard::untransform::avx2::*;
+use dxt_lossless_transform_bc1::transforms::standard::untransform::bench::{
+    avx2::unpck_detransform, *,
+};
 use safe_allocator_api::RawAlloc;
 
 fn bench_unpck(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut RawAlloc) {
